Update submitForm hook to make the key for the form data in the payload dynamic
This follows what is done in V1 and enables the form data in the payload to be named differently from the form name by checking if a submitVariableName hidden field is present in the form and if so then using it's value as the data key rather than the form name.

This gives us the ability to use the form name for other things such as form versioning, which was started to be done in Man Dec but not fully used.