Update submitForm hook to only send businessKey in the payload for forms
It's not necessary to send businessKey in the payload for tasks because the process has already started at this point and the business key is already defined.

This is required by Man Dec because the encryption function for task data is expecting a specific payload in the form of a camunda task object which doesn't include businessKey.