Update FormDetails to display form data correctly in different forms
Update `FormDetails` to display form data correctly in different forms

Previously it was possible to open all the forms for a process in the Case History section of Case View at the same time and doing so would populate all the open forms with the data from the last form that was opened.

This is not fixed so that only one form for a process can be opened at a time and that form displays the correct data.